Follow-up on: Handle induction variables defined outside the loop
Previous commit was not working in case a schedule dimension was not referred
by an access pattern (i.e., the t loop in stencils). Move the logic at the
beginning of the function. The idea to map from isl to pet is the following:
Given a schedule S0[i, j, k] and a loop table which contains C0, C1, C2, map
dimension i with C0, which is the induction variable of the topmost loop. Then
map j with C1 and k with C2.