Add subindex lowering + testsuite
This commit adds support for lowering polygeist.subindex to memref.subview. This PR addressses this as a lowering pass rather than modifying the internals of Polygeist, since it's faster for me, an outsider, to implement :). In reality, I'd suspect memref.subview subsumes the capabilities of polygeist.subindex.

In addition to this, sets up a lit test suite for polygeist-opt and friends allowing for polygeist transformations to be tested.

Can be considered a partial fix to #111 (only considers statically sized memrefs, for now...).