Add a pass converting scf.for(scf.if) to scf.while
The following construct

    scf.for %i ... iter_args(%c = %true) {
      %r = scf.if %c {
        ...
      } else {
        scf.yield %false
      }
      scf.yield %r : i1
    }

is used by Polygeist to represent the C++ break construct.
Unfortunately, lowering it to LLVM IR as is precludes some further
optimizations because, in many cases, it cannot be converted into a
branch to the loop successor block when the break condition triggers,
resulting in spurious loop iterations only incrementing the induction
variable. Add a pass that rewrites this construct into the following
while loop

    scf.while (%i, %c = %true) {
      %end = arith.cmpi slt, %i, upper-bound
      %cond = arith.andi %c, %end
      scf.condition(%cond)
    } do {
      ...
    }

that actually stops the iteration. This only works for perfectly-nested
"scf.if" at the moment.