WIP: alternative lowering for memrefs
Introduce an alternative lowering scheme for memref types. A n-D memref
is always lowered to a pointer to an (n-1)-D nested LLVM array type, a
0-D and a 1-D memref is lowered to a bare pointer. This applies both at
function boundaries and within functions, unlike the MLIR's bare pointer
calling convention. This lowering is only possible for memrefs with
static shapes except for the outermost dimension, which is consistent
with the kinds of nested array structure C allows and therefore suitable
for a C frontend. However, it only supports a subset of memref dialect
operations that are possible to implement using this data structure. For
example, view, subview and transpose cannot be implemented as they
result in non-row-major layouts unexpressible with C nested arrays.

This simplifies the lowering and the generated IR for all supported
memref operations and removes the need for the ABI duplication at the
function boundary: memrefs are lowered to the same LLVM IR as equivalent
nested arrays.