Add missing dependencies from check-polygeist-opt to FileCheck and not
The invocation of the build target `check-polygeist-opt` in a fresh
unified build of LLVM, MLIR, Clang, and Polygeist fails due to the
absence of the test utilities `FileCheck` and `not`.

This commit adds dependencies to the utilities, such that the test
succeeds even when the utilities have not been built manually.